+++ changelog.md
@@ -2,6 +2,7 @@
 
 ## Upcoming
 ### Changes
+* Autumn and Winter no longer prevent grass from spreading; instead, there's a chance for grass to fail to spread
 * Updated biometer texture and recipe, it's now a Sextant. Thanks to [Zeykehasnt](https://github.com/Zeykehasnt) ([#377](https://github.com/Insane96/IguanaTweaksReborn/issues/377))
 * Increased back breed fail chance